Modularized build tools
All build tools are now modularized and they
must implement two functions, `isEligable` and `settings`.
`isEligable` allows the tool to check if it can build or not.
If it can, `settings` will be invoked where build settings should
be returned. If it cannot build, `isEligable` should return false
and the next tool in the chain will be invoked. If no tool can
build, nothing will happen.