Skip error when building on save with no tool
Because buildOnSave is a global configuration a user who
prefers this will have it set. If said user just opens and
save a random file, there will be a no tool build error.
It is unlikely that the user wants to see this error when
building on save without a build tool, so skip it.

Fixes #102