Do things a different way
git commit -F - <<COMMIT
Do things a different way

Don't put the path in quotes.
Replace all POSIX path delimiters with backslash if the platform is Windows.
Likewise, replace all Windows path delimiters with a forward slash if the platform is POSIX.
Also, change the arguments to child_process.spawn so that shCmdArg, exec, and args are each separate.