Remove cmdQuote function from utils.js
Also, if the platform is a POSIX-like platform, join exec and args into one string so it can be executed with sh.
Judging by the way the author wrote the unit tests, sh should only be true if the cmd string is one or more bash/shell commands. sh should be false otherwise (e.g. if cmd refers to an .sh or .bat script).