Fix #484, #447
In TargetManager, getTargets, and by extension fillTargets, has an
implicit requirement that pathTarget.targets is not empty. If it is we
re-call refreshTargets which would end up with infinite recursion. We
don't actually need to call fillTargets as we already know which targets
are available, we can just set them ourselves